apiserver: Rework restore API restrictions to use restrictRoot
This removes a bunch of boilerplate. This is the last of restricted
roots to be refactored.
